pyproject.toml
[project]
name = "adk-sample"
version = "0.1.0"
description = "Google ADKを使用したAIエージェント開発サンプル"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "litellm",
    "a2a-sdk",
    "uvicorn",
    "langchain",
    "langchain-openai",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
